# api-spec.yaml
openapi: 3.0.3
info:
  title: StreamVault API
  description: |
    Professional video hosting platform API. Authenticate with a JWT Bearer token or API key (`sv_live_xxx`).
    
    All timestamps are **UNIX epoch integers** (seconds).
  version: 1.0.0
  contact:
    name: StreamVault Support
    url: https://streamvault.link
servers:
  - url: http://localhost:3000
    description: Local development
  - url: https://api.streamvault.link
    description: Production
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    workspaceId:
      in: header
      name: X-Workspace-Id
      required: true
      schema:
        type: string
...
security:
  - bearerAuth: []
paths:
...
  /api/apikeys:
    get:
      tags: [API Keys]
      summary: List API keys for workspace (Pro+)
      parameters:
        - $ref: '#/components/parameters/workspaceId'
      responses:
        '200':
          description: Key list (no secrets shown)
    post:
      tags: [API Keys]
      summary: Create API key — returned only once (Pro+)
      parameters:
        - $ref: '#/components/parameters/workspaceId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [name]
              properties:
                name: { type: string }
      responses:
        '201':
          description: Key created — save the full key, it will not be shown again
        '403':
          description: Pro plan required
...